Refactor - Move event listeners to their own file
