+++ source/js/eventListeners.js
@@ -1,0 +1,64 @@
+import { chat, event_types, eventSource } from "../../../../../../script.js";
+import { saveMetadataDebounced } from "../../../../../extensions.js";
+import { selected_group } from "../../../../../group-chats.js";
+import { addGroupStatusButtons, fetchStatus, getActiveParticipants, getStatusDepth, log } from "../../index.js";
+import { createCharStatus, getCharStatus } from "./statusControls.js";
+
+// ? event_types.GROUP_UPDATED doesn't matter, status will update when that character sends a message
+/*
+    # TODO
+    - Change of heart, maybe use GROUP_UPDATED for:
+        - [X] Check if the group member exists in metadata
+        - [X] If it doesn't - add it
+    - [ ] Display a table with statuses in the last message of each participant
+*/
+
+export function startListeners() {
+    eventSource.on(event_types.GROUP_UPDATED, async (...args) => {
+        log("GROUP_UPDATED", args);
+
+        for (const char of getActiveParticipants()) {
+            if (!getCharStatus(char)) createCharStatus(char, getStatusDepth(chat, char));
+        }
+
+        saveMetadataDebounced();
+        addGroupStatusButtons();
+    });
+
+    eventSource.on(event_types.CHAT_CHANGED, async (...args) => {
+        log("CHAT_CHANGED", args);
+
+        if (!args[0]) return;
+        if (selected_group) addGroupStatusButtons();
+
+        fetchStatus({forceUIUpdate: true});
+    });
+
+    eventSource.on(event_types.CHARACTER_MESSAGE_RENDERED, async (...args) => {
+        log("CHARACTER_MESSAGE_RENDERED", args);
+        fetchStatus({newMessID: args[0]});
+    });
+
+    eventSource.on(event_types.USER_MESSAGE_RENDERED, async (...args) => {
+        log("USER_MESSAGE_RENDERED", args);
+        fetchStatus({newMessID: args[0]});
+    });
+
+    eventSource.on(event_types.GENERATION_AFTER_COMMANDS, async (...args) => {
+        log("GENERATION_AFTER_COMMANDS", args);
+        fetchStatus({newMessID: chat.length, depthModifier: 1});
+    });
+
+    eventSource.on(event_types.MORE_MESSAGES_LOADED, async (...args) => {
+        log("MORE_MESSAGES_LOADED", args);
+        fetchStatus({forceUIUpdate: true});
+    });
+
+    eventSource.on(event_types.MESSAGE_DELETED, async (...args) => {
+        log("MESSAGE_DELETED", args);
+
+        const id = (args[0] ?? chat.length) - 1;
+
+        fetchStatus({newMessID: id});
+    });
+}
